requirements.txt
Pillow
numpy
streamlit
torch
albumentations
matplotlib
torchvision